Crate canyon_executor

A CodeExecutor specialization which uses natively compiled runtime when the wasm to be executed is equivalent to the natively compiled code.

Structs

Executor
A unit struct which implements NativeExecutionDispatch feeding in the hard-coded runtime.

NativeExecutor
A generic CodeExecutor implementation that uses a delegate to determine wasm code equivalence and dispatch to native code when possible, falling back on WasmExecutor when not.